chore: move execution of multiple files to utils
